+++ app/content/ContentPageClient.jsx
@@ -1,7 +1,7 @@
 "use client";
-import Navbar from "@/components/Navbar";
-import Footer from "@/components/Footer";
-import Newsletter from "@/components/Newsletter";
+import Navbar from "../../components/Navbar";
+import Footer from "../../components/Footer";
+import Newsletter from "../../components/Newsletter";
 import Image from "next/image";
 import { motion, useScroll, useTransform } from "framer-motion";
 import { Share2, Facebook, Twitter, Linkedin, Copy, Play } from "lucide-react";
