+++ app/layout.jsx
@@ -1,6 +1,6 @@
 import "./globals.css";
 import { Inter, Playfair_Display } from "next/font/google";
-import { defaultMetadata } from "@/lib/seo";
+import { defaultMetadata } from "../lib/seo";
 
 const inter = Inter({ subsets: ["latin"], variable: "--font-inter" });
 const playfair = Playfair_Display({ subsets: ["latin"], variable: "--font-playfair" });
