+++ app/page.jsx
@@ -1,15 +1,15 @@
-import Navbar from "@/components/Navbar";
-import Hero from "@/components/Hero";
-import SectionHeader from "@/components/SectionHeader";
-import PostCard from "@/components/PostCard";
-import VideoCard from "@/components/VideoCard";
-import Newsletter from "@/components/Newsletter";
-import Footer from "@/components/Footer";
+import Navbar from "../components/Navbar";
+import HeroSection from "../components/HeroSection";
+import SectionHeader from "../components/SectionHeader";
+import PostCard from "../components/PostCard";
+import VideoCard from "../components/VideoCard";
+import Newsletter from "../components/Newsletter";
+import Footer from "../components/Footer";
 import Image from "next/image";
 import { motion } from "framer-motion";
-import { featuredPosts, latestPosts, videos } from "@/lib/data";
+import { featuredPosts, latestPosts, videos } from "../lib/data";
 import HomePageClient from "./HomePageClient";
-import { getPageMetadata } from "@/lib/seo";
+import { getPageMetadata } from "../lib/seo";
 
 export const metadata = getPageMetadata("home");
