+++ components/FramerCard.jsx
