+++ components/HeroSection.jsx
@@ -1,5 +1,3 @@
-import Navbar from './Navbar.jsx';
-
 export default function HeroSection({ title = "Destination" }) {
   return (
     <div
@@ -14,8 +12,6 @@ export default function HeroSection({ title = "Destination" }) {
         backgroundPosition:"center"
       }}
     >
-      {/* Navbar */}
-      <Navbar />
 
       {/* Travel Icon 
       <div style={{
