--- src/app/page.js
@@ -1,10 +1,0 @@
-
-
-export default function Home() {
-  return (
-    <div>
-
-      this is main page js
-    </div>
-  );
-}
